Migrate status endpoints, with integration tests
